Converted as much of the code to Barr-C Standard as possible while resolving all comments thru Doxygen suggestion
