--- test/include/math.h
@@ -1,7 +1,0 @@
-#ifndef MATH_H
-#define MATH_H
-
-int add(int x, int y); 
-
-
-#endif
