--- test/src/math.c
@@ -1,8 +1,0 @@
-#include "math.h"
-#include "debug.h"
-
-int add(int x, int y)
-{
-    DEBUG_PRINT("%d + %d", x, y);
-    return x + y;
-}
